perf(main): Lazy load displayDebugInfo()
