+++ src/main.js
@@ -1,5 +1,4 @@
 const config = require('./config-schema.json');
-const helpers = require('./helpers');
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved
 const { CompositeDisposable } = require('atom');
 
@@ -7,6 +6,7 @@ const { CompositeDisposable } = require('atom');
 let format = null;
 let formatOnSave = null;
 let warnAboutLinterEslintFixOnSave = null;
+let displayDebugInfo = null;
 let subscriptions = null;
 
 // HACK: lazy load most of the code we need for performance
@@ -34,19 +34,13 @@ const lazyWarnAboutLinterEslintFixOnSave = () => {
   warnAboutLinterEslintFixOnSave();
 };
 
-const displayDebugInfo = () => {
-  const info = helpers.getDebugInfo();
-  const details = [
-    `Atom version: ${info.atomVersion}`,
-    `prettier-atom version: ${info.prettierAtomVersion}`,
-    `prettier version: ${info.prettierVersion}`,
-    `prettier-eslint version: ${info.prettierESLintVersion}`,
-    `prettier-atom configuration: ${JSON.stringify(info.prettierAtomConfig, null, 2)}`,
-  ].join('\n');
-  atom.notifications.addInfo('prettier-atom: details on current install', {
-    detail: details,
-    dismissable: true,
-  });
+// HACK: lazy load most of the code we need for performance
+const lazyDisplayDebugInfo = () => {
+  if (!displayDebugInfo) {
+    // eslint-disable-next-line global-require
+    displayDebugInfo = require('./displayDebugInfo');
+  }
+  displayDebugInfo();
 };
 
 // public API
@@ -54,7 +48,7 @@ const activate = () => {
   subscriptions = new CompositeDisposable();
 
   subscriptions.add(atom.commands.add('atom-workspace', 'prettier:format', lazyFormat));
-  subscriptions.add(atom.commands.add('atom-workspace', 'prettier:debug', displayDebugInfo));
+  subscriptions.add(atom.commands.add('atom-workspace', 'prettier:debug', lazyDisplayDebugInfo));
   subscriptions.add(
     atom.workspace.observeTextEditors(editor =>
       subscriptions.add(editor.getBuffer().onWillSave(() => lazyFormatOnSave(editor))),
